Added the logic to player_vs_ai
